requirements.txt
aioconsole==0.8.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
async-timeout==5.0.1
attrs==25.3.0
bech32==1.2.0
cachetools==6.2.0
cffi==2.0.0
cryptography==46.0.1
dnspython==2.8.0
frozenlist==1.7.0
idna==3.10
monstr==0.1.9
multidict==6.6.4
propcache==0.3.2
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
python-dotenv==1.1.1
secp256k1==0.14.0
typing-inspection==0.4.1
typing_extensions==4.15.0
yarl==1.20.1
